seperate syscalls setup to different file
